applied cr remarks
