+++ src/bundle/Controller/UserSettingsController.php
@@ -26,7 +26,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class UserSettingsController extends Controller implements RestrictedControllerInterface
+final class UserSettingsController extends Controller implements RestrictedControllerInterface
 {
     use AuthenticatedRememberedCheckTrait;
